添加工程页面

+++ User/screen/screenStatus.h
@@ -68,7 +68,7 @@ typedef struct{
 	uint16_t change_max_time;						//换气次数最大值
 	uint8_t screen_light_value;						//屏幕亮度调节 
 	uint8_t modbus_address;							//modbus节点地址
-	uint8_t modbus_tran_rate;						//modbus通信速率
+	uint32_t modbus_tran_rate;						//modbus通信速率
 	uint8_t menu_language;							//菜单语言
 	uint32_t flash_this_address;
 	uint32_t flash_write_cnt;
